// src/types/wish.ts
export namespace Adhoc {
    // api 交互时统一为小写，大写字母形式显示在 url 中很怪
    export type CurrencyType = 'sanity' | 'hintpoint';
}

export namespace Wish {
    export type ErrorRes = { status: 'error'; message: string; title: string };
    export type InfoRes = { status: 'info'; message: string; title: string };
    export type SuccessRes = { status: 'success'; message?: string; title?: string };
    export type NormalRes = ErrorRes | InfoRes | SuccessRes;

    export namespace Game {
// ...
        export type PuzzleGroupInfo = { name: string; puzzles: PuzzleInfo[] };
// ...
    }
// ...
    export namespace Puzzle {
        export type SubmitAnswerApi = {
            request: {
                endpoint: 'puzzle/submit_answer';
                payload: { puzzle_key: string; content: string };
            };
            response: NormalRes & { need_reload?: boolean; redirect?: string };
        };

        export type SubmissionRecordData = {
            idx: number;
            team_name: string;
            user_name: string;
            origin: string;
            cleaned: string;
            status: string;
            info: string;
            timestamp_s: number;
        };

        export type GetSubmissionsApi = {
            request: {
                endpoint: 'puzzle/get_submissions';
                payload: { puzzle_key: string };
            };
            response: ErrorRes | (SuccessRes & { submissions: SubmissionRecordData[] });
        };

        export type PuzzleActionData =
            | { type: 'webpage'; name: string; url: string; noreferrer?: boolean }
            | { type: 'media'; name: string; media_url: string };

        export type ClipboardData = { type: string; content: string };

        export type PuzzleGroupInfo = Game.PuzzleGroupInfo;

        export type PuzzleDetailData = {
            key: string;
            title: string;
            return: string;
            desc: string;
            actions: PuzzleActionData[];
            clipboard?: ClipboardData[];
            status: string;
            puzzle_list: PuzzleGroupInfo[];
            unlock_ts?: number;
            pass_ts?: number;
            cold_down_ts?: number;
            special_list?: object;
            correct_answers?: string[];
            answer_display?: string;
            area_name: string;
        };

        export type GetDetailApi = {
            request: {
                endpoint: 'puzzle/get_detail';
                payload: { puzzle_key: string };
            };
            response: ErrorRes | (SuccessRes & { data: PuzzleDetailData });
        };

        export type HintPrice = { type: Adhoc.CurrencyType; price: number };

        export type HintItem = {
            id: number;
            question: string;
            answer: string;
            type: string;
            price: HintPrice[];
            unlock: boolean;
            effective_after_ts: number;
        };

        export type HintRecordData = { list: HintItem[] };

        export type GetHintsApi = {
            request: {
                endpoint: 'puzzle/get_hints';
                payload: { puzzle_key: string };
            };
            response: ErrorRes | (SuccessRes & { data: HintRecordData });
        };

        export type BuyHintApi = {
            request: {
                endpoint: 'puzzle/buy_hint';
                payload: { puzzle_key: string; hint_id: number };
            };
            response: NormalRes;
        };

        export type PuzzleApis = SubmitAnswerApi | GetSubmissionsApi | GetDetailApi | GetHintsApi | BuyHintApi;
    }
// ...
}